Fixed some syntax errors
The random ; at the start of app.txt and rent.txt are because sqlite is
detecting a sequence of unicode characters at the start of those files,
even though there aren't any. It still throws the syntax error at the
ghost unicode characters, but it doesn't interfere with any of the
INSERT INTO statements, the database is unaffected and still works fine.

I tried a bunch of other things to fix it and it is still detecting
characters like a U inside another U and other weird characters I've
never seen before, hence this "band-aid" fix.